requirements.txt
# VegaExchange Backend Dependencies

# Web framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# Database
asyncpg>=0.29.0
psycopg2-binary>=2.9.9  # For psql compatibility if needed

# Data validation
pydantic>=2.5.0

# Environment
python-dotenv>=1.0.0

# Authentication (for future Google OAuth)
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
httpx>=0.26.0  # For OAuth callbacks

# WebSocket support (built into FastAPI/Starlette)
websockets>=12.0

# Utilities
python-multipart>=0.0.6  # For form data
orjson>=3.9.0  # Fast JSON parsing

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
httpx>=0.26.0  # For testing

# Type hints
typing-extensions>=4.9.0
